rename to `LegacyTabs`
