+++ packages/Overview.mdx
@@ -38,7 +38,7 @@ V1 er klar når følgende komponenter er markert som "✅ Felles":
 | Search                                              | 🚸 Ikke påbegynt | Figma - Search                                                                                                                                                | [Github - Search](https://github.com/digdir/designsystem/issues/88)         |
 | [Switch](/docs/felles-switch--docs)                 | ✅ Felles        | [Figma - Switch](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?type=design&node-id=17755-6024&mode=design&t=ztPUyfbDSQjlpEzv-0)        | [Github - Switch](https://github.com/digdir/designsystem/issues/89)         |
 | [Table](/docs/altinn-table--docs)                   | 🔵 Altinn        | Figma - Table                                                                                                                                                 | [Github - Table](https://github.com/digdir/designsystem/issues/90)          |
-| [Tabs](/docs/altinn-tabs--docs)                     | 🔵 Altinn        | [Figma - Tabs](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?type=design&node-id=9551%3A54208&t=Rlfq5UyNZBL69dFr-1)                    | [Github - Tabs](https://github.com/digdir/designsystem/issues/91)           |
+| [Tabs](/docs/avviklet-legacytabs--docs)             | 🔵 Altinn        | [Figma - Tabs](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?type=design&node-id=9551%3A54208&t=Rlfq5UyNZBL69dFr-1)                    | [Github - Tabs](https://github.com/digdir/designsystem/issues/91)           |
 | [Tag](/docs/felles-tag--docs)                       | ✅ Felles        | [Figma - Tag](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?node-id=10185%3A59053&t=7Q2N4sUdQGhFZrPh-1)                                | [Github - Tag](https://github.com/digdir/designsystem/issues/322)           |
 | [Textarea](/docs/felles-textarea--docs)             | ✅ Felles        | [Figma - Text area](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?node-id=6632%3A21873&t=7Q2N4sUdQGhFZrPh-1)                           | [Github - Textarea](https://github.com/digdir/designsystem/issues/323)      |
 | [Textfield](/docs/felles-textfield--docs)           | ✅ Felles        | [Figma - Text Field](https://www.figma.com/file/vpM9dqqQPHqU6ogfKp5tlr/Felles-komponenter?node-id=6632%3A22228&t=7Q2N4sUdQGhFZrPh-1)                          | [Github - Textfield](https://github.com/digdir/designsystem/issues/92)      |
