+++ packages/react/src/components/legacy/LegacyTabs/Tabs.module.css
