+++ packages/react/src/components/legacy/LegacyTabs/index.ts
@@ -1,0 +1,2 @@
+export { LegacyTabs } from './Tabs';
+export type { LegacyTabsProps, LegacyTabItem } from './Tabs';
