--- packages/react/src/components/legacy/Tabs/index.ts
@@ -1,2 +1,0 @@
-export { Tabs } from './Tabs';
-export type { TabsProps, TabItem } from './Tabs';
